+++ src/vue/pages/polls-all/PollViewer.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="poll-viewer">
-    <tabs>
+    <tabs v-if="isLoaded">
       <tab
         :name="'poll-viewer.vote-tab' | globalize"
         id="poll-viewer__poll-vote-form"
@@ -17,6 +17,29 @@
       >
         <poll-attributes :poll="poll" />
       </tab>
+
+      <tab
+        v-if="(isPollOwner() || isPollResultProvider()) && poll.isOpen"
+        :name="'poll-viewer.manage-tab' | globalize"
+        id="poll-viewer__poll-manage-form"
+      >
+        <poll-manage-form
+          :poll="poll"
+          :is-poll-owner="isPollOwner()"
+          :is-poll-result-provider="isPollResultProvider()"
+          @end-time-updated="updatedEndTimeFormSubmitted()"
+          @poll-closed="$emit(EVENTS.pollClosed)"
+          @poll-canceled="$emit(EVENTS.pollCanceled)"
+        />
+      </tab>
+
+      <tab
+        v-if="isPollOwner()"
+        :name="'poll-viewer.participants-tab' | globalize"
+        id="poll-viewer__poll-participants"
+      >
+        <poll-participants :poll="poll" />
+      </tab>
     </tabs>
   </div>
 </template>
@@ -27,11 +50,20 @@ import Tab from '@/vue/common/tabs/Tab'
 
 import { PollRecord } from '@/js/records/entities/poll.record'
 
-import PollVoteForm from './PollVoteForm.vue'
-import PollAttributes from './PollAttributes.vue'
+import PollVoteForm from './PollVoteForm'
+import PollAttributes from './PollAttributes'
+import PollManageForm from './PollManageForm'
+import PollParticipants from './PollParticipants'
+import { api } from '@/api'
+import { ErrorHandler } from '@/js/helpers/error-handler'
+import { mapGetters } from 'vuex'
+import { vuexTypes } from '@/vuex'
 
 const EVENTS = {
-  close: 'close',
+  closeDrawer: 'close-drawer',
+  pollClosed: 'poll-closed',
+  pollCanceled: 'poll-canceled',
+  endTimeUpdated: 'end-time-updated',
 }
 
 export default {
@@ -42,18 +74,64 @@ export default {
     Tab,
     PollVoteForm,
     PollAttributes,
+    PollManageForm,
+    PollParticipants,
   },
 
   props: {
-    poll: {
-      type: PollRecord,
+    pollId: {
+      type: String,
       required: true,
     },
   },
 
+  data () {
+    return {
+      poll: {},
+      isLoaded: false,
+      EVENTS,
+    }
+  },
+
+  computed: {
+    ...mapGetters({
+      accountId: vuexTypes.accountId,
+    }),
+  },
+
+  async created () {
+    await this.loadPoll()
+    this.isLoaded = true
+  },
+
   methods: {
+    async loadPoll () {
+      const endpoint = `/v3/polls/${this.pollId}`
+      try {
+        const { data } = await api.getWithSignature(endpoint, {
+          include: ['participation', 'participation.votes'],
+        })
+        this.poll = new PollRecord(data)
+      } catch (error) {
+        ErrorHandler.processWithoutFeedback(error)
+      }
+    },
+
     voteFormSubmitted (event) {
-      this.$emit(EVENTS.close)
+      this.$emit(EVENTS.closeDrawer)
+    },
+
+    async updatedEndTimeFormSubmitted () {
+      this.$emit(EVENTS.endTimeUpdated)
+      await this.loadPoll()
+    },
+
+    isPollOwner () {
+      return this.accountId === this.poll.ownerId
+    },
+
+    isPollResultProvider () {
+      return this.accountId === this.poll.resultProviderId
     },
   },
 }