+++ src/vue/pages/polls/PollRequests.vue
@@ -16,7 +16,6 @@ export default {
   components: {
     SubmoduleImporter,
   },
-
   data: _ => ({
     PollRequestsModule,
   }),